Fairly workable split between packages
