+++ frontend/packages/rotations/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./rotations";
 export * from "./math";
+export * from "./geometry";
